+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed bug with text values longer than 256 characters and `_all` field - see [#850](https://github.com/ankane/searchkick/issues/850)
 - Fixed issue with `_all` field in `searchable`
+- Fixed `exclude` option with `word_start`
 
 ## 2.1.1